+++ terraform/hydrocron-dynamo.tf
@@ -51,6 +51,33 @@ resource "aws_dynamodb_table" "hydrocron-swot-node-table" {
   }
 }
 
+resource "aws_dynamodb_table" "hydrocron-swot-prior-lake-table" {
+  name         = "hydrocron-swot-prior-lake-table"
+  billing_mode = "PAY_PER_REQUEST"
+  hash_key     = "lake_id"
+  range_key    = "range_start_time"
+  attribute {
+    name = "lake_id"
+    type = "S"
+  }
+  attribute {
+    name = "range_start_time"
+    type = "S"
+  }
+  attribute {
+    name = "granuleUR"
+    type = "S"
+  }
+
+  global_secondary_index {
+    name               = "GranuleURIndex"
+    hash_key           = "granuleUR"
+    range_key          = "range_start_time"
+    projection_type    = "INCLUDE"
+    non_key_attributes = ["lake_id", "collection_shortname", "collection_version", "crid", "cycle_id", "pass_id", "continent_id", "ingest_time"]
+  }
+}
+
 resource "aws_dynamodb_table" "hydrocron-track-ingest-table" {
   name         = "hydrocron-track-ingest-table"
   billing_mode = "PAY_PER_REQUEST"